Moving some scripts around and adding a second clss to handle the buttons for the DogBrowser
