+++ Assets/Prefabs/Pickup Pup/DogBrowser.prefab
@@ -426,6 +426,7 @@ GameObject:
   - component: {fileID: 114251177231270808}
   - component: {fileID: 114930641940952366}
   - component: {fileID: 114829477178275386}
+  - component: {fileID: 114275684136827656}
   m_Layer: 5
   m_Name: DogBrowser
   m_TagString: Untagged
@@ -1389,6 +1390,24 @@ MonoBehaviour:
   m_FillAmount: 1
   m_FillClockwise: 1
   m_FillOrigin: 0
+--- !u!114 &114275684136827656
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1425115051427350}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3e8624e755bd94e2496c600707395c24, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  alternateSprites: []
+  pageButtonRef: {fileID: 1160778752883856}
+  pageButtonParent: {fileID: 224220737125569122}
+  pageBackwardButton: {fileID: 114112114804278772}
+  pageForwardButton: {fileID: 114164380738010282}
+  closeWindowHitArea: {fileID: 114292257243731584}
+  rehomeButton: {fileID: 114945358630142426}
 --- !u!114 &114286731513834564
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -2811,13 +2830,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   alternateSprites: []
-  pageButtonRef: {fileID: 1160778752883856}
-  pageButtonParent: {fileID: 224220737125569122}
   defaultStartPage: 0
-  pageBackwardButton: {fileID: 114112114804278772}
-  pageForwardButton: {fileID: 114164380738010282}
-  closeWindowHitArea: {fileID: 114292257243731584}
-  rehomeButton: {fileID: 114945358630142426}
 --- !u!114 &114837713938564970
 MonoBehaviour:
   m_ObjectHideFlags: 1
