+++ Assets/Scripts/UI/Pickup Pup/AdoptionScreen.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 23befe6040159418e85c7408f41681e4
+folderAsset: yes
+timeCreated: 1483914440
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
