+++ Assets/Scripts/UI/Pickup Pup/Core.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 6d2062593b2ff4753bde95cbf9c784e8
+folderAsset: yes
+timeCreated: 1483914346
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
