+++ Assets/Scripts/UI/Pickup Pup/Core/PPUIButtonController.cs
@@ -1,0 +1,10 @@
+﻿/*
+ * Author(s): Isaiah Mann
+ * Description: A UI script for handles the buttons of a particular screen
+ * Usage: Should be a owned by a UI Screen or panele
+ */
+
+public class PPUIButtonController : PPUIElement 
+{	
+
+}
