+++ Assets/Scripts/UI/Pickup Pup/Currency.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 540d6de4c59074ae1bab634ad7ba3ad0
+folderAsset: yes
+timeCreated: 1483914410
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
