+++ Assets/Scripts/UI/Pickup Pup/Currency/CurrencyDisplay.cs
