+++ Assets/Scripts/UI/Pickup Pup/DogBrowser.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: dcc34c0e1603f43be858cea121c4d8be
+folderAsset: yes
+timeCreated: 1483914466
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
