+++ Assets/Scripts/UI/Pickup Pup/DogBrowser/DogBrowserButtonController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3e8624e755bd94e2496c600707395c24
+timeCreated: 1483914484
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
