+++ Assets/Scripts/UI/Pickup Pup/DogSlot.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 4051c30c0360b471b91e1a1a4e6735b7
+folderAsset: yes
+timeCreated: 1483914331
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
