+++ Assets/Scripts/UI/Pickup Pup/DogSlot/DogAdoptionSlot.cs
