+++ Assets/Scripts/UI/Pickup Pup/DogSlot/DogOutsideSlot.cs.meta
