+++ Assets/Scripts/UI/Pickup Pup/DogSlot/DogSlot.cs
