+++ Assets/Scripts/UI/Pickup Pup/StartScreen.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 93e54f4e1adc24dcf81f5b8c045bd01f
+folderAsset: yes
+timeCreated: 1483914451
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
